+++ src/auth/index.ts
@@ -1,3 +1,3 @@
 export { default as AuthClient } from "./authClient";
 export { default as authProvider } from "./authProvider";
-export * from './authType';
+export * from "./authType";